Updated names for files and directories
